refactoring out the singleton into a util file
Adding GUI message for ARP and incorrect password

Fixed minor bug with double sudoing causing GUI to hangup in permissions
file
